Moved KMinValues to use byte array
After some profiling it seemed that most of the computations were spent
converting from a byte array (from leveldb) to an integer array for
KMinValues.  I rewrote KMinValues to operator on a byte array in order
to fix this!

A caveat is that now we reverted the KMinValue hashes to be uint64 in
order to speed up byte-wise comparisons.